Copy multiple request responses
Select the request responses to be copied from the
history. Open context menu and navigate to Extensions,
Copy Request Response and choose an option.

The requests and response are copied in the same order
they are displayed in the history. I would rather have
them always sorted by number or request sent date, but
this seems not possible since the number and timingData
is not available. So change your sorting in the history
table first before copying. What also works is to move
request in the order to be copied to the Organizer first.